Modificando o diagrama de classe, acrescentando o que faltava. Fazendo modificações na classe cadastro, alterando os valores para unsigned lon long para a integração. Modificando o atmega, mas isso ficará para questões de teste integrado, mudança para o atmega2560.

+++ Cadastro.cpp
@@ -13,13 +13,13 @@ Cadastro::Cadastro() {
 
 Cadastro::~Cadastro() {}
 
-void Cadastro::adiciona(unsigned long matricula){
+void Cadastro::adiciona(unsigned long long matricula){
 
 	_cadastrados.push(matricula);
 
 }
 
-bool Cadastro::procura(unsigned long matricula){
+bool Cadastro::procura(unsigned long long matricula){
 
 	int in = _cadastrados.procura(matricula);
 
@@ -36,7 +36,7 @@ bool Cadastro::procura(unsigned long matricula){
 
 }
 
-bool Cadastro::exclui(unsigned long matricula){
+bool Cadastro::exclui(unsigned long long matricula){
 	return _cadastrados.exclui(matricula);
 
 }
